create chat_completion.py
